+++ packages/client/src/features/game/utils.ts
@@ -1,4 +1,4 @@
-import { Card } from '../types'
+import { Card } from './types'
 
 /**
  * Создание колоды карт
