--- packages/client/src/features/gameCanvas/index.ts
@@ -1,1 +1,0 @@
-export { default as GameCanvas } from './ui/gameCanvas'
