--- packages/client/src/features/gameSlice/index.ts
@@ -1,1 +1,0 @@
-export { default as gameSlice } from './ui/gameSlice'
