+++ packages/client/src/pages/finish/index.ts
@@ -1,1 +1,1 @@
-export { FinishPage } from './ui/finish'
+export { FinishPage } from './view'
