+++ packages/client/src/pages/game/index.ts
@@ -1,1 +1,1 @@
-export { GamePage } from './ui/game'
+export { GamePage } from './view'
