--- packages/client/src/shared/types/index.ts
@@ -1,5 +1,0 @@
-export interface Card {
-  suit: string
-  value: string
-  hidden?: boolean
-}
